+++ src/globalMsa.cpp
@@ -53,7 +53,7 @@ void dinara::globalMsa(
     // Compute the maximum length of the input sequences.
     uint64_t maxLength = 0;
     for(const auto& p: sequences) {
-        maxLength = max(maxLength, uint64_t(p.first.size()));
+        maxLength = max(maxLength, p.first.size());
     }
 
     // If short enough, use spoa.
@@ -423,7 +423,7 @@ bool dinara::globalMsaSpoa(
     if(sequences.size() > 1) {
         uint64_t maxLength = 0;
         for(const auto& sequence: sequences) {
-            maxLength = max(maxLength, uint64_t(sequence.first.size()));
+            maxLength = max(maxLength, sequence.first.size());
         }
         if(maxLength > maximumMsaLength) {
             return false;
